migrated methods for the producer into a separate file for extensibility... will work to clean it up. I'm also handling recursive js objects incorrectly...
