+++ bin/puppeteer-producer.js
@@ -3,6 +3,7 @@ const puppeteer = require('puppeteer');
 
 const setup = require('./puppeteer-setup');
 const args = require('./puppeteer-args').argv(process.argv, 1);
+const harness = require('./qunit-harness');
 
 if (!args) {
   console.log("Usage: node run-qunit-chrome.js options");
@@ -36,6 +37,8 @@ ipc.serve(() => {
           var args = Array.isArray(arg) ? arg : [].slice.call(arguments);
           ipc.server.emit(socket, 'results', JSON.stringify(args));
         };
+        //TODO - need to work on making this a bit more succinct... move this code to another file?
+
 
         //TODO need to ensure that this is working properly
         // This injects the script tags to the page prior to it this is injected into the frame,
@@ -43,118 +46,14 @@ ipc.serve(() => {
         sendMessage('inject', options.inject);
         await scripts.forEach(async script => page.addScriptTag(script)); // TODO need to test if this works...
 
-        // Attach to browser console log events, and log to node console
-        await page.on('console', (...params) => {
-          for (let i = 0; i < params.length; ++i) {
-            console.log(`${params[i].text}`);
-            sendMessage('console', params[i]); //TODO should I only log 'log' level
-          }
-          //Logs everything here indiscriminately, , the params object is constructed as such:
-          /*
-            ConsoleMessage(
-              type: 'log' | 'error' | etc,
-              text: 'the outpur of the console message'
-              args: 'a list of the arguments'
-            )
-          */
-        });
+
+        //This abstracts out the event handling syntax to try to make this more extensible in the future... lets see if this works.
+        await harness.setupPageEvents(page, sendMessage);
 
         var moduleErrors = [];
         var testErrors = [];
         var assertionErrors = [];
-
-        await page.on('error', err => {
-          console.error(err);
-          console.error(err.stack);
-          sendMessage('error.onError', err, err.stack, "node error"); // this logs an error thrown by node.
-        });
-
-        await page.on('pageerror', err => {
-          sendMessage('error.onError', err);
-        });
-
-        await page.on('request', req => {
-          sendMessage('requestIssued', req);
-          //TODO - this is a really messy bit of code... will have to determine the best solution at a later point.
-        });
-
-        await page.on('requestfailed', req => {
-          sendMessage('resourceFailed', req);
-        });
-
-        await page.on('requestfinished', req => {
-          sendMessage('resourceFinished', req);
-        });
-
-        await page.on('response', res => {
-          sendMessage('responseReceived', res);
-        });
-
-        await page.exposeFunction('harness_moduleDone', context => {
-          if (context.failed) {
-            var msg = "Module Failed: " + context.name + "\n" + testErrors.join("\n");
-            moduleErrors.push(msg);
-            testErrors = [];
-          }
-        });
-
-        await page.exposeFunction('harness_testDone', context => {
-          if (context.failed) {
-            var msg = "  Test Failed: " + context.name + assertionErrors.join("    ");
-            testErrors.push(msg);
-            assertionErrors = [];
-            process.stdout.write("F");
-            sendMessage('F');
-          } else {
-            process.stdout.write(".");
-            sendMessage('.');
-          }
-        });
-
-        await page.exposeFunction('harness_log', context => {
-          if (context.result) {
-            return;
-          } // If success don't log
-
-          var msg = "\n    Assertion Failed:";
-          if (context.message) {
-            msg += " " + context.message;
-          }
-
-          if (context.expected) {
-            msg += "\n      Expected: " + context.expected + ", Actual: " + context.actual;
-          }
-
-          assertionErrors.push(msg);
-          sendMessage(msg);
-        });
-
-        await page.exposeFunction('harness_done', context => {
-          console.log("\n");
-
-          if (moduleErrors.length > 0) {
-            for (var idx = 0; idx < moduleErrors.length; idx++) {
-              console.error(moduleErrors[idx] + "\n");
-            }
-          }
-
-          var stats = [
-            "Time: " + context.runtime + "ms",
-            "Total: " + context.total,
-            "Passed: " + context.passed,
-            "Failed: " + context.failed
-          ];
-          console.log(stats.join(", "));
-
-          browser.close();
-          if (context.failed > 0) {
-            ipc.emit(socket, 'finish');
-            process.exit(1);
-          } else {
-            ipc.emit(socket, 'finish');
-            process.exit();
-          }
-        });
+        await harness.setupExposedMethods(page, sendMessage, moduleErrors, testErrors, assertionErrors);
 
         await page.goto("file://" + pageUrl);
 
@@ -188,8 +87,11 @@ ipc.serve(() => {
         browser.close();
       })
       .catch((error) => {
-        console.log(error);
+        ipc.serve.emit(socket, "error".error, error);
+
+        console.error(error);
       });
+
   });
   ipc.server.on('socket.disconnected', (data, socket) => {
     console.log("DISCONNECTED\n\n", arguments);
